+++ scripts/shared/settings.js
@@ -26,5 +26,6 @@ var STORAGE_KEYS = {
     likesResetTime: 'likes_reset_time',
     likesCount: 'previous_likes_count',
     visitedProfiles: 'visited_profiles',
-    interestedProfileMap: 'interested_profile_map'
+    interestedProfileMap: 'interested_profile_map',
+    likedByUserIds: 'cupid_liked_by_user_ids'
 };